requirements.txt
Flask==2.3.2
mysql_connector_repackaged==0.3.1
python-dotenv==1.0.0
